+++ modules/utils.py
@@ -13,11 +13,6 @@ import yaml
 DIR_PATH = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 NAGIOS_PATH = "/usr/lib/nagios/plugins/"
 
-# Redis Keys
-VALID_HOSTS = "host_names"
-HOST_STATUS = "host_status"
-COMMAND_TASK_ID = "command_task_id"
-
 # status description
 SERVICE_STATUSES = ["OK", "Warning", "Critical", "Unknown"]
 
@@ -54,21 +49,6 @@ def read_yaml(file):
         logging.error(f"Error parsing YAML file {file}")
 
     return result
-
-
-# read a key from the database, converting to dict
-def read_db(db, db_key):
-    result = {}
-
-    if(db.exists(db_key)):
-        result = json.loads(db.get(db_key))
-
-    return result
-
-
-# write a value to the datase, converting to JSON string
-def write_db(db, db_key, db_value):
-    db.set(db_key, json.dumps(db_value))
 
 
 # write JSON to file